Collection and TestTool moved to testutils
