+++ sdl-core/src/test/scala/io/smartdatalake/util/spark/dataset/TransformTest.scala
@@ -19,9 +19,9 @@
 
 package io.smartdatalake.util.spark.dataset
 
-import io.smartdatalake.util.TestTool
+import io.smartdatalake.testutils.TestTool
 import io.smartdatalake.util.spark.GetSession.{createSparkSession, loggEnv}
-import io.smartdatalake.util.spark.dataset.Collection._
+import io.smartdatalake.testutils.spark.dataset.Collection._
 import org.apache.spark.sql.functions.{col, lit}
 import org.apache.spark.sql.types._
 import org.apache.spark.sql.{Row, SparkSession}
